[POL-17] Сделал новые Entity, начал переписывать DAO и репозитории

+++ app/src/main/java/com/example/polycareer/domain/model/AnswersResponse.kt
@@ -1,3 +1,3 @@
 package com.example.polycareer.domain.model
 
-data class AnswersResponse (val result: Result, val answers: List<Int>?)
+data class AnswersResponse (val result: Result, val answers: List<Long>?)
